Dev lorenz

+++ app/Http/Controllers/Admin/DishController.php
@@ -32,7 +32,6 @@ class DishController extends Controller
     ];
     public function index()
     {
-
         $dishes = Dish::where('user_id', auth()->id())->get();
         return view('admin.dishes.index', compact('dishes'));
     }
@@ -131,6 +130,9 @@ class DishController extends Controller
      */
     public function destroy(Dish $dish)
     {
+        if ($dish->user_id !== auth()->id()) {
+            return redirect()->route('admin.dishes.index')->with('error', 'You are not authorized to delete this dish.');
+        }
         $dish->delete();
         return redirect()->route('admin.dishes.index')->with('success', 'Dish deleted!');
     }
